Made a simple working node for the torpedo task
