+++ vision_layer/vision_tasks/CMakeLists.txt
@@ -38,9 +38,9 @@ target_link_libraries(base_class ${catkin_LIBRARIES} ${OpenCV_LIBRARIES} vision_
 # add_dependencies( line_task base_class ${vision_tasks_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 # target_link_libraries(line_task base_class ${catkin_LIBRARIES} ${OpenCV_LIBRARIES} vision_commons )
 
-# add_library(torpedo_task SHARED src/torpedo.cpp include/torpedo.h include/base_class.h)
-# add_dependencies(base_class torpedo_task ${vision_tasks_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
-# target_link_libraries(base_class torpedo_task ${catkin_LIBRARIES} ${OpenCV_LIBRARIES} vision_commons )
+add_library(torpedo_task SHARED src/torpedo.cpp include/torpedo.h include/base_class.h)
+add_dependencies(torpedo_task base_class ${vision_tasks_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
+target_link_libraries(torpedo_task base_class ${catkin_LIBRARIES} ${OpenCV_LIBRARIES} vision_commons )
 
 add_library(gate_task SHARED src/gate.cpp include/gate.h include/base_class.h)
 add_dependencies(gate_task base_class  ${vision_tasks_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
@@ -71,4 +71,5 @@ target_link_libraries(vision_node
   gate_task 
   path_marker
   marker
+  torpedo_task
 )
